# docs/swagger.yaml
definitions:
...
  dto.Buyer:
    properties:
      address:
        type: string
      buyerID:
        type: string
      cart:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
      city:
        type: string
      name:
        type: string
      password:
        type: string
      payment:
        type: string
      phoneNumber:
        type: string
      province:
        type: string
      surname:
        type: string
      transaction:
        items:
          $ref: '#/definitions/dto.Transaction'
        type: array
      username:
        type: string
      zip:
        type: string
    type: object
  dto.BuyerPaymentRequest:
    properties:
      amount:
        type: number
      buyerID:
        type: string
      createdAt:
        type: string
      paymentMethod:
        type: string
    required:
    - amount
    - buyerID
    - createdAt
    - paymentMethod
    type: object
  dto.BuyerRegisterRequest:
    properties:
      address:
        type: string
      city:
        type: string
      name:
        type: string
      password:
        type: string
      payment:
        type: string
      phoneNumber:
        type: string
      province:
        type: string
      surname:
        type: string
      username:
        type: string
      zip:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dto.Product:
    properties:
      amount:
        type: integer
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      imageURL:
        type: string
      price:
        type: number
      productID:
        type: string
      productName:
        type: string
      sellerID:
        type: string
      tag:
        items:
          type: string
        type: array
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  dto.Transaction:
    properties:
      amount:
        minimum: 0
        type: number
      data:
        type: string
      paymentMethod:
        type: string
      product:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
    type: object
  dto.UpdateCartRequest:
    properties:
      product:
        $ref: '#/definitions/dto.Product'
    type: object
  model.Buyer:
    properties:
      address:
        type: string
      buyerID:
        type: string
      cart:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
      city:
        type: string
      name:
        type: string
      password:
        type: string
      payment:
        type: string
      phoneNumber:
        type: string
      province:
        type: string
      surname:
        type: string
      transaction:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
      username:
        type: string
      zip:
        type: string
    type: object
...
  model.Transaction:
    properties:
      amount:
        minimum: 0
        type: number
      data:
        type: string
      paymentMethod:
        type: string
      product:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
    type: object
info:
  contact: {}
paths:
...
  /buyer/:
    get:
      consumes:
      - application/json
      description: Retrieves all buyers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Buyer'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all buyers
      tags:
      - buyer
    post:
      consumes:
      - application/json
      description: Creates a new buyer in the database
      parameters:
      - description: Buyer to create
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/dto.BuyerRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Buyer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new buyer
      tags:
      - buyer
  /buyer/{buyer_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a buyer's data by their ID
      parameters:
      - description: Buyer ID
        in: path
        name: buyer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Buyer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a buyer by ID
      tags:
      - buyer
  /buyer/{buyer_id}/cart:
    patch:
      consumes:
      - application/json
      description: Adds the product if not in the cart, removes it if already in the
        cart
      parameters:
      - description: Buyer ID
        in: path
        name: buyer_id
        required: true
        type: string
      - description: Product ID to add/remove
        in: body
        name: product_id
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a product in buyer's cart
      tags:
      - buyer
  /buyer/{buyer_id}/payment:
    post:
      consumes:
      - application/json
      description: Request payment for buyer and redirect the buyer to payment
      parameters:
      - description: buyer payment request
        in: body
        name: buyerPaymentRequest
        required: true
        schema:
          $ref: '#/definitions/dto.BuyerPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Request payment for buyer
      tags:
      - buyer
  /buyer/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing buyer's data by their ID
      parameters:
      - description: Buyer ID
        in: path
        name: id
        required: true
        type: string
      - description: Buyer data to update
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/model.Buyer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Buyer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a buyer by ID
      tags:
      - buyer
...
swagger: "2.0"